Create README.md
Designed an application using Kotlin programming language that gives a customer to order provided food items from home. It also designed with many features such as adding items to wishlist, cart, ordering them. The items that are depicted created from the "Postman" which acts like a database and sends the request to internet, then receive the required or mentioned data to the app